chore(725): Meged Advisor subcategory with Chats
