+++ GUI/src/RootComponent.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import {Navigate, Route, Routes} from 'react-router-dom'
 import { Layout } from './components'
-import AdvisorsPage from './pages/AdvisorsPage'
 import ChatsPage from './pages/ChatsPage'
 import FeedbackPage from './pages/FeedbackPage'
 import NotFoundPage from './pages/NotFoundPage'
@@ -19,7 +18,6 @@ const RootComponent: React.FC = () => {
         <Route path={ROUTES.CHATS_ROUTE} element={<ChatsPage />} />
         {/* <Route path={ROUTES.BUROKRATT_ROUTE} element={<BurokrattPage />} /> To be Added in release 2.1 or later */}
         <Route path={ROUTES.FEEDBACK_ROUTE} element={<FeedbackPage />} />
-        <Route path={ROUTES.ADVISORS_ROUTE} element={<AdvisorsPage />} />
         <Route path={ROUTES.REPORTS_ROUTE} element={<ReportsPage />} />
         <Route path="*" element={<NotFoundPage />} />
       </Route>
