+++ GUI/src/pages/ChatsPage/options.tsx
@@ -47,4 +47,34 @@ export const chatOptions: Option[] = [
     labelKey: 'chats.idle',
     unit: t('units.chats') ?? 'chats',
   },
+  {
+    id: 'chat_forwards',
+    labelKey: 'chats.chat_forwards',
+    subOptions: [
+      { id: 'receivedChats', labelKey: 'chats.forwards.from_csa', color: '#FFB511' },
+      { id: 'forwardedChats', labelKey: 'chats.forwards.to_csa', color: '#ED7D31' },
+      { id: 'forwardedExternally', labelKey: 'chats.forwards.to_other', color: '#8BB4D5' },
+    ],
+    unit: t('units.chats') ?? 'chats',
+  },
+  {
+    id: 'avg_pick_time',
+    labelKey: 'chats.avg_pick_time',
+    unit: t('units.minutes') ?? 'minutes',
+  },
+  {
+    id: 'avg_present_csa',
+    labelKey: 'chats.avg_present_csa',
+    unit: t('units.counselors') ?? 'counselors',
+  },
+  {
+    id: 'num_chats_csa',
+    labelKey: 'chats.num_chats_csa',
+    unit: t('units.chats') ?? 'chats',
+  },
+  {
+    id: 'avg_chat_time_csa',
+    labelKey: 'chats.avg_chat_time_csa',
+    unit: t('units.minutes') ?? 'minutes',
+  },
 ];
